refactor: 💡 reorganize API namespaces and unify controller structure
- Renamed DTOs `RegisterUserInput` and `RegisterUserResponse` to `Model` subnamespace for consistency
- Moved `RegistrationController` and `EmailVerificationController` to `Controller/API/Auth/Register/`
- Moved `ResetPassword*Controller` to `Controller/API/Auth/ResetPassword/`
- Updated all controller references in ApiResources to match new namespaces
- Extracted image endpoint from `MoviesEntryPoint` into dedicated `ImageEntryPoint`
- Renamed `MovieImageOutput` to `ImageOutput` and moved to `Image/Model` namespace
- Renamed and relocated `MovieImageController` to `ImageController` under `Controller/API/Shared/`
- Moved `PersonController` and `MovieControllers` to new API module layout (`API/Content/`)
