+++ src/ApiResource/Auth/Register/RegisterEntryPoint.php
@@ -4,14 +4,15 @@ namespace App\ApiResource\Auth\Register;
 
 use ApiPlatform\Metadata\ApiResource;
 use ApiPlatform\Metadata\Post;
-use App\Controller\Auth\Register\RegistrationController;
+use App\ApiResource\Auth\Register\Model\RegisterUserInput;
+use App\ApiResource\Auth\Register\Model\RegisterUserResponse;
 
 #[ApiResource(
     operations: [
         new Post(
             uriTemplate: '/register',
             status: 201,
-            controller: RegistrationController::class,
+            controller: \App\Controller\API\Auth\Register\RegistrationController::class,
             input: RegisterUserInput::class,
             output: RegisterUserResponse::class,
             name: 'post_register',
