+++ src/ApiResource/Auth/ResetPassword/ResetPasswordEntryPoint.php
@@ -3,12 +3,8 @@
 namespace App\ApiResource\Auth\ResetPassword;
 
 use ApiPlatform\Metadata\ApiResource;
-use App\Controller\Auth\ResetPassword\ResetPasswordRequestController;
-use App\Controller\Auth\ResetPassword\ResetPasswordResetController;
 use ApiPlatform\Metadata\Post;
 use ApiPlatform\Metadata\Get;
-use App\ApiResource\Auth\ResetPassword\Model\Request;
-use App\ApiResource\Auth\ResetPassword\Model\Reset;
 use App\ApiResource\Auth\ResetPassword\Model\Reset\ValidateResetTokenResponse;
 use App\ApiResource\Auth\ResetPassword\Model\Request\RequestPasswordResetInput;
 use App\ApiResource\Auth\ResetPassword\Model\Request\RequestPasswordResetOutput;
@@ -20,7 +16,7 @@ use App\ApiResource\Auth\ResetPassword\Model\Reset\ResetPasswordOutput;
         new Post(
             uriTemplate: '/reset-password/request',
             status: 200,
-            controller: ResetPasswordRequestController::class,
+            controller: \App\Controller\API\Auth\ResetPassword\ResetPasswordRequestController::class,
             input: RequestPasswordResetInput::class,
             output: RequestPasswordResetOutput::class,
             name: 'post_reset_password_request'
@@ -28,7 +24,7 @@ use App\ApiResource\Auth\ResetPassword\Model\Reset\ResetPasswordOutput;
         new Get(
             uriTemplate: '/reset-password/validate/{token}',
             status: 200,
-            controller: ResetPasswordResetController::class,
+            controller: \App\Controller\API\Auth\ResetPassword\ResetPasswordResetController::class,
             output: ValidateResetTokenResponse::class,
             read: false,
             name: 'get_validate_reset_token'
@@ -36,7 +32,7 @@ use App\ApiResource\Auth\ResetPassword\Model\Reset\ResetPasswordOutput;
         new Post(
             uriTemplate: '/reset-password/reset',
             status: 200,
-            controller: ResetPasswordResetController::class,
+            controller: \App\Controller\API\Auth\ResetPassword\ResetPasswordResetController::class,
             input: ResetPasswordInput::class,
             output: ResetPasswordOutput::class,
             name: 'post_reset_password'
