+++ src/ApiResource/Movies/MoviesEntryPoint.php
@@ -4,15 +4,13 @@ namespace App\ApiResource\Movies;
 
 use ApiPlatform\Metadata\ApiResource;
 use ApiPlatform\Metadata\Get;
-use App\Controller\Movies\MovieDetailsController;
 use App\ApiResource\Movies\Details\MovieDetailsOutput;
-use App\Controller\Movies\SearchMoviesController;
 use App\ApiResource\Movies\Search\SearchMoviesOutput;
-use App\Controller\Movies\MovieImageController;
-use App\ApiResource\Movies\Image\MovieImageOutput;
 use Symfony\Component\Routing\Requirement\Requirement;
 use ApiPlatform\OpenApi\Model\Operation;
 use ApiPlatform\OpenApi\Model\Parameter;
+use App\Controller\API\Content\Movies\SearchMoviesController;
+use App\Controller\API\Content\Movies\MovieDetailsController;
 
 #[ApiResource(
     operations: [
@@ -48,38 +46,6 @@ use ApiPlatform\OpenApi\Model\Parameter;
             ),
             output: SearchMoviesOutput::class,
             name: 'get_movies_search'
-        ),
-        new Get(
-            uriTemplate: '/movies/image/{size}/{endpoint}',
-            controller: MovieImageController::class,
-            openapi: new Operation(
-                parameters: [
-                    (new Parameter(
-                        name: 'endpoint',
-                        in: 'path',
-                        description: 'Identificador do recurso de imagem',
-                        required: true,
-                        schema: ['type' => 'string',]
-                    )),
-                    (new Parameter(
-                        name: 'size',
-                        in: 'path',
-                        description: 'Tamanho da imagem (opcional)',
-                        required: false,
-                        schema: [
-                            'type' => 'string',
-                            'enum' => [
-                                'w92','w154','w185','w342','w500','w780',
-                                'w300','w780','w45','w185','h632','original',
-                            ],
-                            'default' => 'original',
-                            'required' => true,
-                        ]
-                    )),
-                ],
-            ),
-            output: MovieImageOutput::class,
-            name: 'get_movies_img'
         ),
     ],
 )]
